Unify single- and multi-neuron task code
