+++ obi_one/scientific/tasks/em_synapse_mapping/config.py
@@ -1,6 +1,5 @@
 import logging
 from enum import StrEnum
-from pathlib import Path
 from typing import ClassVar
 
 from entitysdk.client import Client
@@ -8,7 +7,6 @@ from entitysdk.models import Entity
 from entitysdk.types import EntityType, TaskActivityType, TaskConfigType
 from pydantic import Field
 
-from obi_one.core.base import OBIBaseModel
 from obi_one.core.block import Block
 from obi_one.core.schema import SchemaKey, UIElement
 from obi_one.core.single import SingleConfigMixin
@@ -24,7 +22,7 @@ class BlockGroup(StrEnum):
 
 
 class EMSynapseMappingScanConfig(InfoScanConfig):
-    """Map location of afferent synapses from EM onto a spiny morphology."""
+    """Map location of afferent synapses from EM onto one or more spiny morphologies."""
 
     name: ClassVar[str] = "Map synapse locations"
 
@@ -43,13 +41,15 @@ class EMSynapseMappingScanConfig(InfoScanConfig):
     )
 
     def input_entities(self, db_client: Client) -> list[Entity]:
-        return [self.initialize.spiny_neuron.entity(db_client=db_client)]
+        return [n.entity(db_client=db_client) for n in self.initialize.neurons]
 
     class Initialize(Block):
-        spiny_neuron: CellMorphologyFromID = Field(  # | MEModelFromID
-            title="EM skeletonized morphology",
-            description="""A neuron morphology with spines obtained from an electron-microscopy
-            datasets through the skeletonization task.""",
+        # We use a tuple instead of a list to avoid getting it taken as scan dimensions
+        # in the scan config.
+        neurons: tuple[CellMorphologyFromID | MEModelFromID, ...] = Field(
+            title="Neurons",
+            description="Neurons to include in the circuit (>= 1).",
+            min_length=1,
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.MODEL_IDENTIFIER,
                 SchemaKey.ENTITY_QUERY: {
@@ -60,31 +60,34 @@ class EMSynapseMappingScanConfig(InfoScanConfig):
                 },
             },
         )
-        edge_population_name: str = Field(
-            default="afferent_synapses",
-            min_length=1,
-            title="Edge population name",
-            description="Name of the edge population to write the synapse information into",
+        physical_edge_population_name: str = Field(
+            title="Physical edge population name",
+            description="Edge population for connections between neurons in the set.",
+            default="physical_connections",
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.STRING_INPUT,
             },
         )
-        node_population_pre: str = Field(
-            default="afferent_neurons",
-            min_length=1,
-            title="Presynaptic node population name",
-            description="""Name of the node population to write the information about the
-            innervating neurons into""",
+        virtual_edge_population_name: str = Field(
+            title="Virtual edge population name",
+            description="Edge population for connections from virtual neurons.",
+            default="virtual_afferents",
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.STRING_INPUT,
             },
         )
-        node_population_post: str = Field(
-            default="biophysical_neuron",
-            min_length=1,
-            title="Postsynaptic node population name",
-            description="""Name of the node population to write the information about the
-            synaptome neuron into""",
+        biophysical_node_population: str = Field(
+            title="Biophysical node population name",
+            description="Node population for the physical neurons in the circuit.",
+            default="biophysical_neurons",
+            json_schema_extra={
+                SchemaKey.UI_ELEMENT: UIElement.STRING_INPUT,
+            },
+        )
+        virtual_node_population: str = Field(
+            title="Virtual node population name",
+            description="Node population for external presynaptic neurons.",
+            default="virtual_afferent_neurons",
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.STRING_INPUT,
             },
@@ -106,48 +109,3 @@ class EMSynapseMappingSingleConfig(EMSynapseMappingScanConfig, SingleConfigMixin
     _single_task_activity_type: ClassVar[TaskActivityType] = (
         TaskActivityType.em_synapse_mapping__execution
     )
-
-
-class EMSynapseMappingMultipleConfig(OBIBaseModel):
-    """Config for multi-neuron synapse mapping."""
-
-    name: ClassVar[str] = "Map synapse locations (multi-neuron)"
-    description: ClassVar[str] = (
-        "Map afferent synapses from EM onto multiple spiny morphologies, "
-        "producing a circuit with physical internal connections and virtual external inputs."
-    )
-    coordinate_output_root: Path = Field(title="Output directory")
-
-    class Initialize(Block):
-        # We use a tuple instead of a list to avoid getting it taken as scan dimensions in the
-        # scan config.
-        neurons: tuple[CellMorphologyFromID | MEModelFromID, ...] = Field(
-            title="Neurons",
-            description="Neurons to include in the multi-neuron circuit.",
-            min_length=2,
-        )
-        physical_edge_population_name: str = Field(
-            title="Physical edge population name",
-            description="Edge population for connections between neurons in the set.",
-            default="physical_connections",
-        )
-        virtual_edge_population_name: str = Field(
-            title="Virtual edge population name",
-            description="Edge population for connections from virtual neurons.",
-            default="virtual_afferents",
-        )
-        biophysical_node_population: str = Field(
-            title="Biophysical node population name",
-            description="Node population for the physical neurons in the circuit.",
-            default="biophysical_neurons",
-        )
-        virtual_node_population: str = Field(
-            title="Virtual node population name",
-            description="Node population for external presynaptic neurons.",
-            default="virtual_afferent_neurons",
-        )
-
-    initialize: Initialize
-
-    def enforce_no_multi_param(self) -> None:
-        """Override: allow list-valued neurons field."""
