+++ obi_one/scientific/tasks/em_synapse_mapping/publication_links.py
@@ -2,7 +2,7 @@ import logging
 
 from entitysdk import Client
 from entitysdk._server_schemas import (
-    PublicationType,  # NOQA: PLC2701
+    PublicationType,
 )
 from entitysdk.models import (
     EMDenseReconstructionDataset,
